### 2025-05-17 #8213 @editorialbot
[PRE REVIEW]: pyETA: A real-time fixation stream generation and analytical eye-tracking toolbox
**Submitting author:** @VinayIN (<a href="http://orcid.org/0000-0002-6318-3241">Binay Kumar Pradhan</a>)
**Repository:** https://github.com/VinayIN/pyETA-toolbox
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
