### 2025-06-13 #8213 @samhforbes
@ihrke @kinleyid would either of you be interested in reviewing this software for JOSS?
